Implementing SpeakersPage and styling SchedulePage
The SpeakersPage is quite simple for a while. It is just a grid with speakers
cards containing their names and roles. I'm still thinking about a good way to
provide the speakers description and if it is a good idea to have a line with
their main speech title or a catchphrase.

I positioned the day and weekday row, and the speeches container to the center
and added another row to show the day of the event (1º DIA, 2º DIA, etc).

As we don't know how many speeches there will be, the SchedulePage may need
some readjustments, since the container's height increase when adding a new
one. I thought of positioning the v-row with the day and weekday at left, and
the speeches container at right with vertical scrolling in case of having too
many of them.